Create ESG_FACTOR
The idea is simple; firms that have improved the ESG the most are expected to outperform. According to the paper, it is expected that the market could react to a change in rating in a relatively short period. However, the advantages of a better-rated ESG portfolio are expected to be apparent only in the long term, for example, because of increased cash flows, etc. Therefore, the strategy is to overweight, relative to the MSCI World Index, companies that increased their ESG ratings most during the recent past and underweight those with decreased ESG ratings. Where the increases and decreases are based on a 12-month ESG momentum. Last but not least, this paper from the MSCI uses the risk model GEM3S, but the idea is important, not the risk model.